apps/tokenservice/pyproject.toml
[tool.poetry]
name = "token-service"
version = "0.1.0"
description = ""
authors = ["rag-example-team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
azure-identity = "^1.17.1"
pycryptodome = "^3.20.0"
python-jose = {extras = ["pycryptodome"], version = "^3.3.0"}
toml = "^0.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
